requirements.txt
numpy==1.23.5
streamlit==1.32.2
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0